Task 6

+++ pointers_arrays_strings/6-puts2.c
@@ -9,7 +9,7 @@
  */
 void puts2(char *str)
 {
-	long int i = 0;
+	int i = 0;
 
 	while (str[i])
 	{
